+++ HISTORY.md
@@ -4,6 +4,7 @@ History of features added to LiChess Tools in time (reversed order)
   - Interactive lesson Preview mode sticks between page reloads as well
   - Option to open links from move list comments that lead to studies in the same window
   - Transpositions in variation arrows
+  - readgame command
 
 16 Nov
   - Analysis Wiki FEN search